Extract Timer for Supervisor's timed tasks
