+++ test/unit/supervisor_test.rb
@@ -140,7 +140,7 @@ class SupervisorTest < ActiveSupport::TestCase
 
     def assert_registered_supervisor(pid)
       skip_active_record_query_cache do
-        processes = find_processes_registered_as("Supervisor")
+        processes = find_processes_registered_as("Supervisor(forks)")
         assert_equal 1, processes.count
         assert_nil processes.first.supervisor
         assert_equal pid, processes.first.pid
